+++ packages/ra-ui-materialui/src/configurable/Configurable.stories.tsx
@@ -17,7 +17,7 @@ export default {
     title: 'ra-ui-materialui/configurable/Configurable',
 };
 
-const ConfigurableTextBlock = React.forwardRef<
+const TextBlock = React.forwardRef<
     HTMLDivElement,
     { preferencesKey?: string; children?: any }
 >(({ children, preferencesKey }, ref) => {
@@ -70,46 +70,37 @@ const TextBlockEditor = ({ preferencesKey }: { preferencesKey?: string }) => {
     );
 };
 
-const TextBlock = ({ preferencesKey, ...props }: any) => (
+const ConfigurableTextBlock = ({ preferencesKey, ...props }: any) => (
     <Configurable editor={<TextBlockEditor />} preferencesKey={preferencesKey}>
-        <ConfigurableTextBlock {...props} />
+        <TextBlock {...props} />
     </Configurable>
 );
 
-const ConfigurableSalesBlock = React.forwardRef<HTMLDivElement>(
-    (props, ref) => {
-        const [showDate] = useStore('salesBlock.showDate', true);
-        return (
-            <Box
-                display="flex"
-                border="solid 1px lightgrey"
-                borderRadius={3}
-                p={1}
-                margin={1}
-                width={200}
-                ref={ref}
-            >
-                <Box flex="1" mr={1}>
-                    <Typography variant="h6">Sales</Typography>
-                    {showDate && (
-                        <Typography variant="caption">Today</Typography>
-                    )}
-                    <Typography variant="h4" textAlign="right" mt={2}>
-                        $4,452
-                    </Typography>
-                </Box>
-                <Box
-                    bgcolor="lightgrey"
-                    display="flex"
-                    alignItems="center"
-                    p={1}
-                >
-                    <TimelineIcon />
-                </Box>
+const SalesBlock = React.forwardRef<HTMLDivElement>((props, ref) => {
+    const [showDate] = useStore('salesBlock.showDate', true);
+    return (
+        <Box
+            display="flex"
+            border="solid 1px lightgrey"
+            borderRadius={3}
+            p={1}
+            margin={1}
+            width={200}
+            ref={ref}
+        >
+            <Box flex="1" mr={1}>
+                <Typography variant="h6">Sales</Typography>
+                {showDate && <Typography variant="caption">Today</Typography>}
+                <Typography variant="h4" textAlign="right" mt={2}>
+                    $4,452
+                </Typography>
             </Box>
-        );
-    }
-);
+            <Box bgcolor="lightgrey" display="flex" alignItems="center" p={1}>
+                <TimelineIcon />
+            </Box>
+        </Box>
+    );
+});
 
 const SalesBlockEditor = () => {
     const [showDate, setShowDate] = useStore('salesBlock.showDate', true);
@@ -128,9 +119,9 @@ const SalesBlockEditor = () => {
     );
 };
 
-const SalesBlock = props => (
+const ConfigurableSalesBlock = props => (
     <Configurable editor={<SalesBlockEditor />}>
-        <ConfigurableSalesBlock {...props} />
+        <SalesBlock {...props} />
     </Configurable>
 );
 
@@ -140,8 +131,8 @@ export const Basic = () => (
         <InspectorButton />
         <hr />
         <Box display="flex" alignItems="flex-start">
-            <TextBlock />
-            <SalesBlock />
+            <ConfigurableTextBlock />
+            <ConfigurableSalesBlock />
         </Box>
     </PreferencesEditorContextProvider>
 );
@@ -152,9 +143,9 @@ export const Nested = () => (
         <InspectorButton />
         <hr />
         <Box display="flex" alignItems="flex-start">
-            <TextBlock>
-                <SalesBlock />
-            </TextBlock>
+            <ConfigurableTextBlock>
+                <ConfigurableSalesBlock />
+            </ConfigurableTextBlock>
         </Box>
     </PreferencesEditorContextProvider>
 );
@@ -165,8 +156,8 @@ export const MultipleInstances = () => (
         <InspectorButton />
         <hr />
         <Box display="flex" alignItems="flex-start">
-            <TextBlock preferencesKey="foo" />
-            <TextBlock preferencesKey="bar" />
+            <ConfigurableTextBlock preferencesKey="foo" />
+            <ConfigurableTextBlock preferencesKey="bar" />
         </Box>
     </PreferencesEditorContextProvider>
 );
