+++ packages/ra-ui-materialui/src/list/SimpleList/SimpleList.stories.tsx
@@ -1,11 +1,7 @@
 import * as React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import { PreferencesEditorContextProvider, I18nContextProvider } from 'ra-core';
-import polyglotI18nProvider from 'ra-i18n-polyglot';
-import englishMessages from 'ra-language-english';
 
 import { SimpleList } from './SimpleList';
-import { Inspector, InspectorButton } from '../../configurable';
 
 export default { title: 'ra-ui-materialui/list/SimpleList' };
 
@@ -45,24 +41,4 @@ export const Basic = () => (
             tertiaryText={record => record.year}
         />
     </MemoryRouter>
-);
-
-const i18nProvider = polyglotI18nProvider(() => englishMessages, 'en');
-export const Configurable = () => (
-    <I18nContextProvider value={i18nProvider}>
-        <PreferencesEditorContextProvider>
-            <MemoryRouter>
-                <Inspector />
-                <InspectorButton />
-                <SimpleList
-                    resource="books"
-                    data={data}
-                    primaryText={record => record.title}
-                    secondaryText={record => record.author}
-                    tertiaryText={record => record.year}
-                    sx={{ margin: 2 }}
-                />
-            </MemoryRouter>
-        </PreferencesEditorContextProvider>
-    </I18nContextProvider>
 );
